Separate scroll-aware widgets from guiutils.h
