+++ rtgui/controlspotpanel.h
@@ -28,6 +28,7 @@
 #include "threadutils.h"
 #include "toolpanel.h"
 #include "widgets/basic/adjuster.h"
+#include "widgets/basic/smartscroll.h"
 
 class MyExpander;
