+++ rtgui/recentbrowser.cc
@@ -16,7 +16,10 @@
  *  You should have received a copy of the GNU General Public License
  *  along with RawTherapee.  If not, see <https://www.gnu.org/licenses/>.
  */
+
 #include "recentbrowser.h"
+
+#include "guiutils.h"
 #include "multilangmgr.h"
 #include "options.h"
