+++ rtgui/tools/colorappearance.h
@@ -25,6 +25,7 @@
 #include "guiutils.h"
 #include "toolpanel.h"
 #include "widgets/basic/adjuster.h"
+#include "widgets/basic/smartscroll.h"
 
 class DiagonalCurveEditor;
 class CurveEditorGroup;
