+++ rtgui/tools/framing.h
@@ -24,6 +24,7 @@
 #include "guiutils.h"
 #include "toolpanel.h"
 #include "widgets/basic/adjuster.h"
+#include "widgets/basic/smartscroll.h"
 
 #include "rtengine/procevents.h"
