+++ rtgui/widgets/basic/smartscroll.cc
@@ -1,0 +1,361 @@
+/*
+ *  This file is part of RawTherapee.
+ *
+ *  Copyright (c) 2004-2010 Gabor Horvath <hgabor@rawtherapee.com>
+ *
+ *  RawTherapee is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  RawTherapee is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with RawTherapee.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include "smartscroll.h"
+
+#include "options.h"
+#include "rtscalable.h"
+
+#include "rtengine/rtapp.h"
+#include "rtengine/rt_math.h"
+
+/*
+ *
+ * Derived class of some widgets to properly handle the scroll wheel ;
+ * the user has to use the Shift key to be able to change the widget's value,
+ * otherwise the mouse wheel will scroll the editor's tabs content.
+ *
+ */
+MyScrolledWindow::MyScrolledWindow ()
+{
+}
+
+bool MyScrolledWindow::on_scroll_event (GdkEventScroll* event)
+{
+    if (!App::get().options().hideTPVScrollbar) {
+        Gtk::ScrolledWindow::on_scroll_event (event);
+        return true;
+    }
+
+    Glib::RefPtr<Gtk::Adjustment> adjust = get_vadjustment();
+    Gtk::Scrollbar *scroll = get_vscrollbar();
+
+    if (adjust && scroll) {
+        const double upperBound = adjust->get_upper();
+        const double lowerBound = adjust->get_lower();
+        double value = adjust->get_value();
+        double step  = adjust->get_step_increment();
+
+        if (event->direction == GDK_SCROLL_DOWN) {
+            const double value2 = rtengine::min<double>(value + step, upperBound);
+
+            if (value2 != value) {
+                scroll->set_value(value2);
+            }
+        } else if (event->direction == GDK_SCROLL_UP) {
+            const double value2 = rtengine::max<double>(value - step, lowerBound);
+
+            if (value2 != value) {
+                scroll->set_value(value2);
+            }
+        } else if (event->direction == GDK_SCROLL_SMOOTH) {
+            const double value2 = rtengine::LIM<double>(value + event->delta_y * step, lowerBound, upperBound);
+
+            if (value2 != value) {
+                scroll->set_value(value2);
+            }
+        }
+    }
+
+    return true;
+}
+
+void MyScrolledWindow::get_preferred_width_vfunc (int &minimum_width, int &natural_width) const
+{
+    natural_width = minimum_width = RTScalable::scalePixelSize(100);
+}
+
+void MyScrolledWindow::get_preferred_height_vfunc (int &minimum_height, int &natural_height) const
+{
+    natural_height = minimum_height = RTScalable::scalePixelSize(50);
+}
+
+void MyScrolledWindow::get_preferred_height_for_width_vfunc (int width, int &minimum_height, int &natural_height) const
+{
+    natural_height = minimum_height = RTScalable::scalePixelSize(50);
+}
+
+/*
+ *
+ * Derived class of some widgets to properly handle the scroll wheel ;
+ * the user has to use the Shift key to be able to change the widget's value,
+ * otherwise the mouse wheel will scroll the toolbar.
+ *
+ */
+MyScrolledToolbar::MyScrolledToolbar ()
+{
+    set_policy (Gtk::POLICY_EXTERNAL, Gtk::POLICY_NEVER);
+    get_style_context()->add_class("scrollableToolbar");
+
+    // Works fine with Gtk 3.22, but a custom made get_preferred_height had to be created as a workaround
+    // taken from the official Gtk3.22 source code
+    //set_propagate_natural_height(true);
+}
+
+bool MyScrolledToolbar::on_scroll_event (GdkEventScroll* event)
+{
+    Glib::RefPtr<Gtk::Adjustment> adjust = get_hadjustment();
+    Gtk::Scrollbar *scroll = get_hscrollbar();
+
+    if (adjust && scroll) {
+        const double upperBound = adjust->get_upper();
+        const double lowerBound = adjust->get_lower();
+        double value = adjust->get_value();
+        double step  = adjust->get_step_increment() * 2;
+        double value2 = 0.;
+
+//        printf("MyScrolledToolbar::on_scroll_event / delta_x=%.5f, delta_y=%.5f, direction=%d, type=%d, send_event=%d\n",
+//                event->delta_x, event->delta_y, (int)event->direction, (int)event->type, event->send_event);
+
+        if (event->direction == GDK_SCROLL_DOWN) {
+            value2 = rtengine::min<double>(value + step, upperBound);
+            if (value2 != value) {
+                scroll->set_value(value2);
+            }
+        } else if (event->direction == GDK_SCROLL_UP) {
+            value2 = rtengine::max<double>(value - step, lowerBound);
+            if (value2 != value) {
+                scroll->set_value(value2);
+            }
+        } else if (event->direction == GDK_SCROLL_SMOOTH) {
+            if (event->delta_x) {  // if the user use a pad, it can scroll horizontally
+                value2 = rtengine::LIM<double>(value + (event->delta_x > 0 ? 30 : -30), lowerBound, upperBound);
+            } else if (event->delta_y) {
+                value2 = rtengine::LIM<double>(value + (event->delta_y > 0 ? 30 : -30), lowerBound, upperBound);
+            }
+            if (value2 != value) {
+                scroll->set_value(value2);
+            }
+        }
+    }
+
+    return true;
+}
+
+void MyScrolledToolbar::get_preferred_height_vfunc (int &minimumHeight, int &naturalHeight) const
+{
+    int currMinHeight = 0;
+    int currNatHeight = 0;
+    std::vector<const Widget*> childs = get_children();
+    minimumHeight = naturalHeight = 0;
+
+    for (auto child : childs)
+    {
+        if(child->is_visible()) {
+            child->get_preferred_height(currMinHeight, currNatHeight);
+            minimumHeight = rtengine::max(currMinHeight, minimumHeight);
+            naturalHeight = rtengine::max(currNatHeight, naturalHeight);
+        }
+    }
+}
+
+MyComboBoxText::MyComboBoxText (bool has_entry) : Gtk::ComboBoxText(has_entry)
+{
+    minimumWidth = naturalWidth = RTScalable::scalePixelSize(70);
+    Gtk::CellRendererText* cellRenderer = dynamic_cast<Gtk::CellRendererText*>(get_first_cell());
+    cellRenderer->property_ellipsize() = Pango::ELLIPSIZE_MIDDLE;
+    add_events(Gdk::SCROLL_MASK|Gdk::SMOOTH_SCROLL_MASK);
+}
+
+bool MyComboBoxText::on_scroll_event (GdkEventScroll* event)
+{
+
+//    printf("MyComboboxText::on_scroll_event / delta_x=%.5f, delta_y=%.5f, direction=%d, type=%d, send_event=%d\n",
+//            event->delta_x, event->delta_y, (int)event->direction, (int)event->type, event->send_event);
+    // If Shift is pressed, the widget is modified
+    if (event->state & GDK_SHIFT_MASK) {
+        Gtk::ComboBoxText::on_scroll_event(event);
+        return true;
+    }
+
+    // ... otherwise the scroll event is sent back to an upper level
+    return false;
+}
+
+void MyComboBoxText::setPreferredWidth (int minimum_width, int natural_width)
+{
+    if (natural_width == -1 && minimum_width == -1) {
+        naturalWidth = minimumWidth = RTScalable::scalePixelSize(70);
+    } else if (natural_width == -1) {
+        naturalWidth =  minimumWidth = minimum_width;
+    } else if (minimum_width == -1) {
+        naturalWidth = natural_width;
+        minimumWidth = rtengine::max(naturalWidth / 2, RTScalable::scalePixelSize(20));
+        minimumWidth = rtengine::min(naturalWidth, minimumWidth);
+    } else {
+        naturalWidth = natural_width;
+        minimumWidth = minimum_width;
+    }
+}
+
+void MyComboBoxText::get_preferred_width_vfunc (int &minimum_width, int &natural_width) const
+{
+    natural_width = rtengine::max(naturalWidth, RTScalable::scalePixelSize(10));
+    minimum_width = rtengine::max(minimumWidth, RTScalable::scalePixelSize(10));
+}
+
+void MyComboBoxText::get_preferred_width_for_height_vfunc (int height, int &minimum_width, int &natural_width) const
+{
+    natural_width = rtengine::max(naturalWidth, RTScalable::scalePixelSize(10));
+    minimum_width = rtengine::max(minimumWidth, RTScalable::scalePixelSize(10));
+}
+
+
+MyComboBox::MyComboBox ()
+{
+    minimumWidth = naturalWidth = RTScalable::scalePixelSize(70);
+}
+
+bool MyComboBox::on_scroll_event (GdkEventScroll* event)
+{
+
+    // If Shift is pressed, the widget is modified
+    if (event->state & GDK_SHIFT_MASK) {
+        Gtk::ComboBox::on_scroll_event(event);
+        return true;
+    }
+
+    // ... otherwise the scroll event is sent back to an upper level
+    return false;
+}
+
+void MyComboBox::setPreferredWidth (int minimum_width, int natural_width)
+{
+    if (natural_width == -1 && minimum_width == -1) {
+        naturalWidth = minimumWidth = RTScalable::scalePixelSize(70);
+    } else if (natural_width == -1) {
+        naturalWidth =  minimumWidth = minimum_width;
+    } else if (minimum_width == -1) {
+        naturalWidth = natural_width;
+        minimumWidth = rtengine::max(naturalWidth / 2, RTScalable::scalePixelSize(20));
+        minimumWidth = rtengine::min(naturalWidth, minimumWidth);
+    } else {
+        naturalWidth = natural_width;
+        minimumWidth = minimum_width;
+    }
+}
+
+void MyComboBox::get_preferred_width_vfunc (int &minimum_width, int &natural_width) const
+{
+    natural_width = rtengine::max(naturalWidth, RTScalable::scalePixelSize(10));
+    minimum_width = rtengine::max(minimumWidth, RTScalable::scalePixelSize(10));
+}
+
+void MyComboBox::get_preferred_width_for_height_vfunc (int height, int &minimum_width, int &natural_width) const
+{
+    natural_width = rtengine::max(naturalWidth, RTScalable::scalePixelSize(10));
+    minimum_width = rtengine::max(minimumWidth, RTScalable::scalePixelSize(10));
+}
+
+MySpinButton::MySpinButton ()
+{
+    Gtk::Border border;
+    border.set_bottom(0);
+    border.set_top(0);
+    border.set_left(3);
+    border.set_right(3);
+    set_inner_border(border);
+    set_numeric(true);
+    set_wrap(false);
+    set_alignment(Gtk::ALIGN_END);
+    set_update_policy(Gtk::SpinButtonUpdatePolicy::UPDATE_IF_VALID); // Avoid updating text if input is not a numeric
+}
+
+void MySpinButton::updateSize()
+{
+    double vMin, vMax;
+    int maxAbs;
+    unsigned int digits, digits2;
+    unsigned int maxLen;
+
+    get_range(vMin, vMax);
+
+    digits = get_digits();
+    maxAbs = (int)(fmax(fabs(vMin), fabs(vMax)) + 0.000001);
+
+    if (maxAbs == 0) {
+        digits2 = 1;
+    } else {
+        digits2 = (int)(log10(double(maxAbs)) + 0.000001);
+        digits2++;
+    }
+
+    maxLen = digits + digits2 + (vMin < 0 ? 1 : 0) + (digits > 0 ? 1 : 0);
+    set_max_length(maxLen);
+    set_width_chars(maxLen);
+    set_max_width_chars(maxLen);
+}
+
+bool MySpinButton::on_key_press_event (GdkEventKey* event)
+{
+    double vMin, vMax;
+    get_range(vMin, vMax);
+
+    if ((event->keyval >= GDK_KEY_a && event->keyval <= GDK_KEY_z)
+            || (event->keyval >= GDK_KEY_A && event->keyval <= GDK_KEY_Z)
+            || event->keyval == GDK_KEY_equal || event->keyval == GDK_KEY_underscore
+            || event->keyval == GDK_KEY_plus || (event->keyval == GDK_KEY_minus && vMin >= 0)) {
+        return false; // Event is propagated further
+    } else {
+        if (event->keyval == GDK_KEY_comma || event->keyval == GDK_KEY_KP_Decimal) {
+            set_text(get_text() + ".");
+            set_position(get_text().length()); // When setting text, cursor position is reset at text start. Avoiding this with this code
+            return true; // Event is not propagated further
+        }
+
+        return Gtk::SpinButton::on_key_press_event(event); // Event is propagated normally
+    }
+}
+
+bool MySpinButton::on_scroll_event (GdkEventScroll* event)
+{
+    // If Shift is pressed, the widget is modified
+    if (event->state & GDK_SHIFT_MASK) {
+        Gtk::SpinButton::on_scroll_event(event);
+        return true;
+    }
+
+    // ... otherwise the scroll event is sent back to an upper level
+    return false;
+}
+
+bool MyHScale::on_scroll_event (GdkEventScroll* event)
+{
+
+//    printf("MyHScale::on_scroll_event / delta_x=%.5f, delta_y=%.5f, direction=%d, type=%d, send_event=%d\n",
+//            event->delta_x, event->delta_y, (int)event->direction, (int)event->type, event->send_event);
+    // If Shift is pressed, the widget is modified
+    if (event->state & GDK_SHIFT_MASK) {
+        Gtk::Scale::on_scroll_event(event);
+        return true;
+    }
+
+    // ... otherwise the scroll event is sent back to an upper level
+    return false;
+}
+
+bool MyHScale::on_key_press_event (GdkEventKey* event)
+{
+
+    if ( event->string[0] == '+' || event->string[0] == '-' ) {
+        return false;
+    } else {
+        return Gtk::Widget::on_key_press_event(event);
+    }
+}
